Edit styles, add function buttons in module users
