+++ css/style.css
@@ -1,6 +1,6 @@
 QWidget{
 	color: rgb(221, 221, 221);
-	font: 10pt "Segoe UI";
+	font: 12pt "Segoe UI";
 }
 
 QToolTip {
@@ -15,6 +15,14 @@ QToolTip {
 	text-align: left;
 	padding-left: 8px;
 	margin: 0px;
+}
+
+QMessageBox{
+	background-color: #333;
+}
+
+QMessageBox QLabel{
+	color: #aaa;
 }
 
 #styles{
@@ -43,10 +51,7 @@ QToolTip {
 #buttonRight .QPushButton:hover { background-color: rgb(44, 49, 57); border-style: solid; border-radius: 4px; }
 #buttonRight .QPushButton:pressed { background-color: rgb(23, 26, 30); border-style: solid; border-radius: 4px; }
 
-/* #minimizarBtn{
-	background-image: url(images/icons/icon_minimize.png);
-} */
-
+/* Bottom Buttons*/
 #buttonsEnd2 .QPushButton { background-color: rgba(206, 206, 206, 0.288); border: none;  border-radius: 5px; }
 #buttonsEnd2 .QPushButton:hover { background-color: rgb(44, 49, 57); border-style: solid; border-radius: 4px; }
 #buttonsEnd2 .QPushButton:pressed { background-color: rgb(23, 26, 30); border-style: solid; border-radius: 4px; }
@@ -70,18 +75,6 @@ QLineEdit{
 
 QTextEdit{
 	color: #000000;
-}
-
-#msg QMessageBox{
-	background-color: #ff0000;
-}
-
-QMessageBox{
-	background-color: #333;
-}
-
-QMessageBox QLabel{
-	color: #aaa;
 }
 
 /*Pages styles*/
@@ -123,10 +116,7 @@ QLineEdit {
 	border-radius: 5px;
 	border: 2px solid rgb(33, 37, 43);
 	padding-left: 10px;
-	selection-color: rgb(255, 255, 255);
-	selection-background-color: rgb(255, 121, 198);
 }
-
 
 QLineEdit:hover {
 	border: 2px solid rgb(64, 71, 88);
