+++ css/stylescss.py
@@ -10,4 +10,5 @@ def stylesincss():
 	"""
 	with open('css/style.css', 'r') as s:
 		data = s.read()
+		print(data)
 		return data
