+++ modules/functions.py
@@ -44,7 +44,7 @@ def cerrarTicket(Noticket, DescripcionResuelto, CerradoDia):
     return bdsql.tableResueltoWrite(Noticket, DescripcionResuelto, CerradoDia)
 
 
-def setTextCont(self):
+def setTextContUsers(self):
         self.labelTitle.setText("AGREGAR USUARIOS")
         self.labelCedula.setText("Cedula")
         self.labelUsuario.setText("Usuario")
@@ -56,3 +56,5 @@ def setTextCont(self):
         self.buttonCerrar.setText("Cerrar")
         self.buttonAddUsers.setText("Add Users")
         self.buttonEditUsers.setText("Edit Users")
+        self.buttonEditar.setText("Editar")
+        self.buttonEliminar.setText("Eliminar")
